1.1.0-70BX06 - Modified ui/web/uploadCert.php to check if the temp file exists before we unlink it.

+++ BlueOnyx/ui/base-ssl.mod/ui/web/uploadCert.php
@@ -52,12 +52,16 @@ if ($save)
 		        // deal with error
 		        $error = new CceError('huh', 0, 'cert', "[[base-ssl.sslImportError$ret]]");
 		        $errors = array($error);
-			unlink($tmp_cert);
+			if (is_file($temp_cert)) {
+			    unlink($tmp_cert);
+			}
 		    }
 		    else
 		    {
 		        header("Location: /base/ssl/siteSSL.php?group=$group");
-        		unlink($tmp_cert);
+			if (is_file($temp_cert)) {
+			    unlink($tmp_cert);
+			}
 		        $helper->destructor();
 		        exit;
 		    }
